requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.3
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.101.1
fastapi-cache2==0.2.1
flower==2.0.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanize==4.8.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
orjson==3.9.5
passlib==1.7.4
pendulum==2.1.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pyasn1==0.5.0
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.0.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2